Rename the project, update modules, paths and requires.
