+++ config.ru
@@ -1,11 +1,11 @@
 # This config file is injected into the local Docker Compose environment
-# for the purposes of engaging the Gollum::Elasticsearch::Middleware plugin
+# for the purposes of engaging the GollumSearch::Middleware plugin
 # in front of the Gollum app.
 
 $stderr.puts '!!! Rack config engaged !!!'
 
-require_relative 'lib/gollum-elasticsearch.rb'
-use Gollum::Elasticsearch::Middleware
+require_relative 'lib/gollum_search.rb'
+use GollumSearch::Middleware
 
 require 'gollum/app'
 wiki_path = '/app/test/examples/lotr.git'
