--- lib/gollum/elasticsearch.rb
@@ -1,10 +1,0 @@
-require 'gollum/elasticsearch/client'
-require 'gollum/elasticsearch/index'
-require 'gollum/elasticsearch/middleware'
-require 'gollum/elasticsearch/version'
-
-module Gollum
-  module Elasticsearch
-
-  end
-end
