--- lib/gollum/elasticsearch/version.rb
@@ -1,5 +1,0 @@
-module Gollum
-  module Elasticsearch
-    VERSION = '0.1.0'
-  end
-end
