+++ lib/gollum_search.rb
@@ -1,0 +1,8 @@
+require 'gollum_search/version'
+require 'gollum_search/middleware'
+# require 'gollum_search/elasticsearch/client'
+# require 'gollum_search/elasticsearchindex'
+
+module GollumSearch
+
+end
