+++ lib/gollum_search/version.rb
@@ -1,0 +1,3 @@
+module GollumSearch
+  VERSION = '0.1.0'
+end
