+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require 'bundler/setup'
-require 'gollum/elasticsearch'
+require 'gollum_search'
 require 'rack/test'
 
 RSpec.configure do |config|
@@ -7,7 +7,7 @@ RSpec.configure do |config|
   config.include Rack::Test::Methods
 
   def app
-    Gollum::Elasticsearch::Middleware
+    GollumSearch::Middleware
   end
 
   # Enable flags like --only-failures and --next-failure
